requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.9.1
aiosignal==1.3.1
alembic==1.12.0
amqp==5.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.0.0
apache-airflow==2.7.2
apache-airflow-providers-common-sql==1.7.2
apache-airflow-providers-ftp==3.5.2
apache-airflow-providers-http==4.5.2
apache-airflow-providers-imap==3.3.2
apache-airflow-providers-sqlite==3.4.3
apispec==6.3.0
appdirs==1.4.4
argcomplete==3.1.2
asgiref==3.7.2
async-timeout==4.0.3
asyncssh==2.21.0
atpublic==6.0.1
attrs==23.1.0
Babel==2.13.0
backoff==1.10.0
billiard==4.2.1
blinker==1.6.3
cachelib==0.9.0
cachetools==5.5.2
cattrs==23.1.2
celery==5.5.2
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
cloudpickle==3.1.1
colorama==0.4.6
colorlog==4.8.0
configobj==5.0.9
ConfigUpdater==3.1.1
connexion==2.14.2
contourpy==1.3.2
cron-descriptor==1.4.0
croniter==1.4.1
cryptography==41.0.4
cycler==0.12.1
databricks-sdk==0.52.0
Deprecated==1.2.14
dictdiffer==0.9.0
dill==0.3.1.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.4.2
docker==7.1.0
docutils==0.20.1
dpath==2.2.0
dulwich==0.22.8
dvc==3.59.2
dvc-data==3.16.10
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
email-validator==1.3.1
entrypoints==0.4
exceptiongroup==1.1.3
fastapi==0.115.12
filelock==3.18.0
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.58.0
frozenlist==1.4.0
fsspec==2025.3.2
funcy==2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.1
google-re2==1.1
googleapis-common-protos==1.60.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.20.1
greenlet==3.0.0
grpcio==1.59.0
gto==1.7.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
inflection==0.5.1
iterative-telemetry==0.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.5.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.8
kombu==5.5.3
lazy-object-proxy==1.9.0
limits==3.6.0
linkify-it-py==2.0.2
lockfile==0.12.2
Mako==1.2.4
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.10.3
mdit-py-plugins==0.4.0
mdurl==0.1.2
mlflow==2.22.0
mlflow-skinny==2.22.0
multidict==6.0.4
networkx==3.4.2
numpy==2.2.5
omegaconf==2.3.0
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
ordered-set==4.1.0
orjson==3.10.18
packaging==23.2
pandas==2.2.3
pathspec==0.11.2
pendulum==2.1.2
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.3.0
prison==0.2.1
prompt_toolkit==3.0.51
protobuf==4.21.12
psutil==5.9.5
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydot==4.0.0
pygit2==1.18.0
Pygments==2.16.1
pygtrie==2.5.0
PyJWT==2.8.0
pyparsing==3.2.3
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.6.0
rich-argparse==1.3.0
rpds-py==0.10.4
rsa==4.9.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.1
scipy==1.15.3
scmrepo==3.3.11
semver==3.0.4
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.16.0
smmap==5.0.2
sniffio==1.3.0
SQLAlchemy==1.4.49
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
sqltrie==0.11.2
starlette==0.46.2
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.6.0
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.3
typing_extensions==4.8.0
tzdata==2025.2
uc-micro-py==1.0.2
unicodecsv==0.14.1
urllib3==1.26.17
uvicorn==0.34.2
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
zc.lockfile==3.0.post1
zipp==3.17.0
